use s3 hosted static media + start css module refactoring
