+++ client/src/components/RelatedPlaylistsList.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import RelatedPlaylist from './RelatedPlaylist.jsx';
-import tab from '../dist/images/tab.png';
 
 function RelatedPlaylistsList(props) {
   return (
     <div>
       <div className="headerContainer">
-        <div className="headerImage"><img src={tab} className="headerIcon" /> </div>
+        <div className="headerImage"><img src="https://audibly-sb-media.s3-us-west-1.amazonaws.com/icons/tab.png" className="headerIcon" /> </div>
         <div className="headerTitle">Related Playlists</div>
         <div className="headerMore">View all</div>
       </div>
