+++ client/src/components/RelatedSongsList.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import RelatedSong from './RelatedSong.jsx';
-import wave from '../dist/images/wave.png';
 
 function RelatedSongsList(props) {
   return (
     <div>
       <div className="headerContainer">
-        <div className="headerImage"><img src={wave} className="headerIcon"/> </div>
+        <div className="headerImage"><img src="https://audibly-sb-media.s3-us-west-1.amazonaws.com/icons/wave.png" className="headerIcon"/> </div>
         <div className="headerTitle">Related Tracks</div>
         <div className="headerMore">View all</div>
       </div>
