+++ package.json
@@ -44,13 +44,13 @@
     "eslint-plugin-react": "^7.19.0",
     "eslint-plugin-react-hooks": "^1.7.0",
     "faker": "^4.1.0",
-    "file-loader": "^6.0.0",
     "jest": "^25.1.0",
     "jest-environment-enzyme": "^7.1.2",
     "jest-enzyme": "^7.1.2",
     "mongodb": "^3.5.5",
     "mongoose": "^5.9.4",
     "nodemon": "^2.0.2",
+    "react-css-modules": "^4.7.11",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11"
   }
